+++ profiles/preview/resourcegraph/mgmt/resourcegraph/models.go
@@ -8,7 +8,7 @@
 
 package resourcegraph
 
-import original "github.com/Azure/azure-sdk-for-go/services/resourcegraph/mgmt/2019-04-01/resourcegraph"
+import original "github.com/Azure/azure-sdk-for-go/services/resourcegraph/mgmt/2021-03-01/resourcegraph"
 
 const (
 	DefaultBaseURI = original.DefaultBaseURI